+++ src/cilabo/labo/developing/fan2021/CommandLineArgs.java
@@ -1,4 +1,4 @@
-package cilabo.labo.develop;
+package cilabo.labo.developing.fan2021;
 
 import java.io.File;
